refactor miniball system so that icons are always updated
