+++ samples/golf/src/golf/GolfState.hpp
@@ -40,6 +40,7 @@ source distribution.
 #include "CPUGolfer.hpp"
 #include "TerrainDepthmap.hpp"
 #include "TerrainChunks.hpp"
+#include "MinimapZoom.hpp"
 #include "server/ServerPacketData.hpp"
 
 #include <crogine/core/State.hpp>
@@ -378,23 +379,7 @@ private:
 
     void updateMiniMap();
 
-    struct MinimapZoom final
-    {
-        glm::vec2 pan = glm::vec2(0.f);
-        float tilt = 0.f;
-        float zoom = 1.f;
-
-        glm::mat4 invTx = glm::mat4(1.f);
-        std::uint32_t shaderID = 0u;
-        std::int32_t matrixUniformID = -1;
-        std::int32_t featherUniformID = -1;
-
-        glm::vec2 mapScale = glm::vec2(0.f); //this is the size of the texture used in relation to world map, ie pixels per metre
-        glm::vec2 textureSize = glm::vec2(1.f);
-
-        void updateShader();
-        glm::vec2 toMapCoords(glm::vec3 worldPos) const;
-    }m_minimapZoom;
+    MinimapZoom m_minimapZoom;
     void retargetMinimap(bool reset);
 
     cro::Entity m_greenCam;
